+++ Pandora/Paradigm/Inventory/Optics.hs
@@ -10,11 +10,13 @@ import Pandora.Pattern.Functor.Divariant ((>->))
 import Pandora.Pattern.Functor.Invariant (Invariant ((<$<)))
 import Pandora.Pattern.Object.Setoid (Setoid ((==)))
 import Pandora.Paradigm.Controlflow.Effect.Interpreted (run, (||=))
+import Pandora.Paradigm.Primary.Functor.Maybe (Maybe)
 import Pandora.Paradigm.Primary.Functor.Product (Product ((:*:)))
 import Pandora.Paradigm.Primary.Transformer.Flip (Flip)
 import Pandora.Paradigm.Primary.Object.Boolean ((?))
 import Pandora.Paradigm.Inventory.Store (Store (Store), position, look, retrofit)
 import Pandora.Paradigm.Schemes.PQ_ (PQ_ (PQ_))
+import Pandora.Paradigm.Schemes.P_T (P_T)
 
 infixr 0 :-.
 infixr 0 :~.
@@ -49,3 +51,8 @@ over lens f = extract . retrofit f . run lens
 -- | Representable based lens
 represent :: (Representable t, Setoid (Representation t)) => Representation t -> t a :-. a
 represent r = PQ_ $ \x -> Store $ r <#> x :*: \new -> tabulate (\r' -> r' == r ? new $ r' <#> x)
+
+type Prism = PQ_ (->) (P_T Store Maybe)
+
+preview :: Prism src tgt -> src -> Maybe tgt
+preview prism = position . run . run prism